Add pixel distribution stuff
Makes a plot of vis6, vis8, and nir distributions for a few different
points. Point is to show that in vis6 and vis8 the distribution is
strongly dominated by low-value pixels, which motivates us looking at
the median/mean of low values for the ground value. The nir plots show
a completely different distribution.

Second plot shows the statistics of vis6 and vis8. I think it's clear
that for land pixels, we can use for the ground value either the
median value of the entire distribution (since it is not skewed by the
few large values) or the mean of the distribution not including those
values larger than the Otsu threshold.