Add hadley cell caption. Stop hyperref. See note
For some reason, the citation \cite{geomar6557} is causing issues. It
seems to be a problem with hyperref struggling to wrap it
appropriately. If you add / remove words around the citation, it will
work or not work depending on how many words there are (and how TeX
wraps the words).

Bit shit, but the only way I can stop it happening is by using the
[draft] option for hyperref.